Move `ChannelUpdaterMock` to separate file and make it internal
